moved make_default_crypto_signer lib to test/module/..
