+++ test/framework/integration_framework/fake_peer/fake_peer.cpp
@@ -15,8 +15,6 @@
 #include "consensus/yac/yac_crypto_provider.hpp"
 #include "cryptography/crypto_provider/crypto_signer.hpp"
 #include "cryptography/default_hash_provider.hpp"
-#include "cryptography/ed25519_sha3_impl/crypto_provider.hpp"
-#include "framework/integration_framework/default_crypto_signer.hpp"
 #include "framework/integration_framework/fake_peer/behaviour/behaviour.hpp"
 #include "framework/integration_framework/fake_peer/block_storage.hpp"
 #include "framework/integration_framework/fake_peer/network/loader_grpc.hpp"
@@ -32,6 +30,7 @@
 #include "logger/logger_manager.hpp"
 #include "main/server_runner.hpp"
 #include "module/shared_model/cryptography/crypto_defaults.hpp"
+#include "module/shared_model/cryptography/make_default_crypto_signer.hpp"
 #include "multi_sig_transactions/transport/mst_transport_grpc.hpp"
 #include "network/impl/async_grpc_client.hpp"
 #include "network/impl/grpc_channel_builder.hpp"
